feat: Better error handling
