+++ src/App/App.tsx
@@ -2,7 +2,7 @@ import { css } from '@emotion/css';
 import { type AppRootProps, type GrafanaTheme2 } from '@grafana/data';
 import { locationService } from '@grafana/runtime';
 import { ErrorBoundary, useStyles2 } from '@grafana/ui';
-import React, { createContext, useEffect, useState } from 'react';
+import React, { createContext, useState } from 'react';
 
 import { type DataTrail } from 'DataTrail';
 import { initFaro } from 'tracking/faro/faro';
@@ -10,6 +10,7 @@ import { getUrlForTrail, newMetricsTrail } from 'utils';
 
 import { ErrorView } from './ErrorView';
 import { AppRoutes } from './Routes';
+import { useCatchExceptions } from './useCatchExceptions';
 import { PluginPropsContext } from '../utils/utils.plugin';
 initFaro();
 
@@ -31,25 +32,7 @@ function App(props: AppRootProps) {
     setTrail(trail);
   };
 
-  const [error, setError] = useState<Error>();
-
-  useEffect(() => {
-    // even though we wrap the app in an ErrorBoundary, some errors are not caught,
-    // so we have to set a global onerror handler to catch these (e.g. error thrown from some click handlers)
-    const onError = (errorEvent: ErrorEvent) => {
-      setError(errorEvent.error);
-    };
-    const onUnHandledRejection = (event: PromiseRejectionEvent) => {
-      setError(new Error(event.reason, { cause: { type: event.type } }));
-    };
-
-    window.addEventListener('error', onError);
-    window.addEventListener('unhandledrejection', onUnHandledRejection);
-    return () => {
-      window.removeEventListener('unhandledrejection', onUnHandledRejection);
-      window.removeEventListener('error', onError);
-    };
-  }, []);
+  const [error, setError] = useCatchExceptions();
 
   if (error) {
     return (
